+++ src/tx-params/create-multisig.js
@@ -15,6 +15,17 @@ import {integerToHexString} from '../utils';
  * @return {TxParams}
  */
 export default function CreateMultisigTxParams({addresses, weights, threshold, feeCoinSymbol, ...otherParams}) {
+    if (addresses.length > 32) {
+        throw new Error('Invalid addresses count, it must not be greater than 32');
+    }
+    if (weights.length !== addresses.length) {
+        throw new Error('Invalid weights count, it must be equal to addresses count');
+    }
+    weights.forEach((weight) => {
+        if (weight > 1023 || weight < 0) {
+            throw new Error('Invalid weight given, it should be between 0 and 1023');
+        }
+    });
     const txData = new TxDataCreateMultisig({
         addresses: addresses.map((address) => toBuffer(address)),
         weights: weights.map((weight) => `0x${integerToHexString(weight)}`),